changed the default timeout to 3 seconds, changed signature of timeout to long
git-svn-id: https://src.springframework.org/svn/spring-osgi/trunk@355 2e8e966d-6415-0410-94b9-dc4e65435d03

+++ spring-osgi-core/src/main/java/org/springframework/osgi/service/OsgiServiceProxyFactoryBean.java
@@ -102,7 +102,7 @@ public class OsgiServiceProxyFactoryBean implements FactoryBean, InitializingBea
 
 	private int contextClassloader;
 
-	private int timeout;
+	private long timeout;
 
 	private PropertyValues properties;
 
@@ -548,7 +548,7 @@ public class OsgiServiceProxyFactoryBean implements FactoryBean, InitializingBea
 	/**
 	 * @return Returns the timeout.
 	 */
-	public int getTimeout() {
+	public long getTimeout() {
 		return timeout;
 	}
 
@@ -556,7 +556,7 @@ public class OsgiServiceProxyFactoryBean implements FactoryBean, InitializingBea
 	/**
 	 * @param timeout The timeout to set.
 	 */
-	public void setTimeout(int timeout) {
+	public void setTimeout(long timeout) {
 		this.timeout = timeout;
 		// This is kind of a hack
 		if (timeout > 0) setRetryTimes((int) (timeout / getRetryDelayMs()));
